recreate pr

+++ mllib/src/main/scala/org/apache/spark/ml/regression/IsotonicRegression.scala
@@ -87,7 +87,7 @@ private[regression] trait IsotonicRegressionBase extends Params with HasFeatures
       col($(featuresCol))
     }
     val w = if (hasWeightCol) {
-      col($(weightCol))
+      col($(weightCol)).cast(DoubleType)
     } else {
       lit(1.0)
     }
